chore: split storage patch into modules
