+++ crates/miden-protocol/src/account/patch/storage/mod.rs
@@ -1,0 +1,14 @@
+mod storage_patch;
+pub use storage_patch::AccountStoragePatch;
+
+mod slot_patch;
+pub use slot_patch::StorageSlotPatch;
+
+mod value_patch;
+pub use value_patch::StorageValuePatch;
+
+mod map_patch;
+pub use map_patch::{StorageMapPatch, StorageMapPatchEntries};
+
+#[cfg(test)]
+mod tests;
